+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Version 0
 
+### 0.5.0
+- Fix in query join interface
+- Add feature to add join conditions for all requests
+
 ### 0.4.0
 - Typescript fixes